requirements.txt
bresenham==0.2.1
cycler==0.11.0
imgaug==0.4.0
kornia==0.5.8
matplotlib==3.1.0
numpy==1.16.4
opencv_contrib_python==3.4.2.17
Pillow==8.4.0
PyYAML==6.0
scipy==1.3.0
tqdm==4.32.1
